+++ backend/src/test/java/com/danielwindel/util/places/PlaceServiceTest.java
@@ -1,4 +1,4 @@
-package com.danielwindel.places;
+package com.danielwindel.util.places;
 
 import com.google.maps.GeoApiContext;
 import com.google.maps.PlaceAutocompleteRequest;
